migrated perf plugin
