+++ x-pack/test/plugin_api_perf/plugins/task_manager_performance/kibana.json
@@ -1,0 +1,9 @@
+{
+  "id": "task_manager_performance",
+  "version": "1.0.0",
+  "kibanaVersion": "kibana",
+  "configPath": ["xpack"],
+  "requiredPlugins": ["taskManager"],
+  "server": true,
+  "ui": false
+}
