+++ x-pack/test/plugin_api_perf/plugins/task_manager_performance/server/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { SampleTaskManagerFixturePlugin } from './plugin';
+
+export const plugin = () => new SampleTaskManagerFixturePlugin();
